restructure commands
