+++ src/modules/osubot/commands/bp.js
@@ -1,46 +1,49 @@
 import util from './_util'
 import { api } from '../web'
 import { userdb } from '../db'
-import { drawRecent } from '../canvas'
+import { drawBest } from '../canvas'
 import MESSAGES from './_messages'
 
 export default {
-    args: '[usr]',
+    args: '<order> [usr]',
     options: util.flatten(util.modes),
     /**
-     * @description Get a user's most recent play
+     * @description Get a user's best performance
      * @param {Message} msg The universal msg object
+     * @param {string} order The username that'll be queried
      * @param {string} usr The username that'll be queried
      * @param {string} mode the mode that will be queried
      */
-    async action(msg, { usr = 'me' }, [ mode = 'o' ]) {
+    async action(msg, { order, usr = 'me' }, [ mode = 'o' ]) {
         mode = util.checkmode(mode)
-        let recent, map, status
+        let best, map, status
+        if (!parseInt(order) || parseInt(order) < 1 || parseInt(order) > 100)
+            msg.send(`osubot: bp: ${MESSAGES.BP_ARGS_FAIL}`)
         if (usr === 'me') {
             try {
                 const doc = await userdb.getByQQ(msg.param.user_id)
                 usr = doc.osuid
             } catch (err) {
-                msg.send(`osubot: stat: ${MESSAGES.QUERY_BIND_FAIL}`)
+                msg.send(`osubot: bp: ${MESSAGES.QUERY_BIND_FAIL}`)
                 return
             }
         }
-        try { 
+        try {
             try {
-                recent = await api.recentQuery({
+                best = (await api.bestQuery({
                     u: usr,
-                    limit: '1',
-                    m: mode,
-                })
+                    limit: order,
+                    m: mode
+                }))[order - 1]
                 ;[map, status] = await Promise.all([
-                    api.mapQuery({ b: recent.beatmap_id }),
+                    api.mapQuery({ b: best.beatmap_id }),
                     api.statQuery({ u: usr }),
                 ])
             } catch (err) {
-                msg.send(`osubot: rec: ${MESSAGES.QUERY_NET_FAIL}`)
+                msg.send(`osubot: bp: ${MESSAGES.QUERY_NET_FAIL}`)
                 return
             }
-            const path = await drawRecent(recent, map, status)
+            const path = await drawBest(best, map, status)
             if (path)
                 msg.send([{
                     type: 'image',
@@ -48,7 +51,7 @@ export default {
                         file: path,
                     }
                 }])
-            else msg.send(`osubot: rec: ${MESSAGES.QUERY_CANVAS_FAIL}`)
+            else msg.send(`osubot: bp: ${MESSAGES.QUERY_CANVAS_FAIL}`)
         } catch (err) {
             msg.error(err)
             return
