+++ src/modules/osubot/commands/index.js
@@ -1,0 +1,9 @@
+import avatar from './avatar'
+import bind from './bind'
+import bp from './bp'
+import db from './db'
+import rec from './rec'
+import roll from './roll'
+import stat from './stat'
+import unbind from './unbind'
+export default { avatar, bind, bp, db, rec, roll, stat, unbind }
